+++ spring-boot-2-demo-app/src/main/java/io/github/mfvanek/spring/boot2/test/controllers/RedirectController.java
@@ -1,4 +1,4 @@
-package io.github.mfvanek.spring.test.controllers;
+package io.github.mfvanek.spring.boot2.test.controllers;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
